requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Data processing
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# File handling
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# ML and AI
scikit-learn==1.3.2
prophet==1.1.4
ollama==0.1.7
sentence-transformers==2.2.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-env==1.1.0
pytest-sugar==0.9.7
pytest-benchmark==4.0.0
pytest-html==4.1.1
httpx==0.25.2
faker==20.1.0
factory-boy==3.3.0
freezegun==1.2.2
responses==0.24.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production
gunicorn==21.2.0
